feat: add embedding Tauri commands and wire run_embedding_refresh option
- Add `embedding_refresh_run` and `embedding_status` Tauri commands to
  commands.rs, backed by `crate::embeddings::service::{refresh_embeddings,
  embedding_status}`.
- Register both commands in lib.rs `run()` builder and
  `generate_typescript_bindings` test.
- Add `run_embedding_refresh: bool` field to `JiraIngestionOptions`
  (defaults false); documents the intended hook path for the command layer
  to trigger a refresh after ingestion without threading SecretStore into
  the ingestion service itself.
- Change `EmbeddingRunSummary.dimension` from `usize` to `u32` so specta
  can emit a TypeScript-safe numeric type (BigInt-free).
- Regenerate src/bindings.ts with new command signatures and embedding
  types (`EmbeddingRunOptions`, `EmbeddingRunSummary`, `EmbeddingRunStatus`,
  `EmbeddingStatusSummary`).

+++ src-tauri/src/embeddings/service.rs
@@ -32,7 +32,9 @@ pub struct EmbeddingRunSummary {
     pub skipped: u32,
     pub failed: u32,
     pub model_id: String,
-    pub dimension: usize,
+    /// Embedding vector dimension. Stored as `u32` rather than `usize` so
+    /// specta can emit a TypeScript-safe numeric type without BigInt risk.
+    pub dimension: u32,
     pub safe_error: Option<String>,
 }
 
@@ -128,7 +130,7 @@ pub fn refresh_embeddings_with_provider(
         response.dimension,
         "l2",
     );
-    let dimension = response.dimension;
+    let dimension = response.dimension as u32;
 
     // Write vectors and update status
     match write_embedding_batch(conn, &claimed, &response, now_utc) {
@@ -240,7 +242,7 @@ pub fn refresh_embeddings(
         response.dimension,
         "l2",
     );
-    let dimension = response.dimension;
+    let dimension = response.dimension as u32;
 
     match write_embedding_batch(conn, &claimed, &response, now_utc) {
         Ok(()) => {}
